Refactoring search method packages
