+++ backend-boxboxd/src/main/java/com/kropsz/github/backendboxboxd/service/search/DriverService.java
@@ -1,4 +1,4 @@
-package com.kropsz.github.backendboxboxd.service;
+package com.kropsz.github.backendboxboxd.service.search;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -7,19 +7,19 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.kropsz.github.backendboxboxd.entities.Team;
-import com.kropsz.github.backendboxboxd.repository.TeamRepository;
+import com.kropsz.github.backendboxboxd.entities.Driver;
+import com.kropsz.github.backendboxboxd.repository.DriverRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class TeamService {
+public class DriverService {
     
-    private final TeamRepository teamRepository;
+    private final DriverRepository driverRepository;
 
-    public static Specification<Team> hasProperty(String property, String value) {
-        return (team, cq, cb) -> cb.equal(team.get(property), value);
+     public static Specification<Driver> hasProperty(String property, String value) {
+        return (driver, cq, cb) -> cb.equal(driver.get(property), value);
     }
 
     public static Sort orderByProperty(String property, Sort.Direction direction) {
@@ -33,12 +33,11 @@ public class TeamService {
     }
 
     @Transactional(readOnly = true)
-    public Page<Team> getTeamsByProperty(String property, String value, Pageable pageable) {
+    public Page<Driver> getDriversByProperty(String property, String value, Pageable pageable) {
         if (property == null && value == null) {
-            return teamRepository.findAll(pageable);
+            return driverRepository.findAll(pageable);
         }
-        Specification<Team> spec = hasProperty(property, value);
-        return teamRepository.findAll(spec, pageable);
+        Specification<Driver> spec = hasProperty(property, value);
+        return driverRepository.findAll(spec, pageable);
     }
-
 }
