+++ backend-boxboxd/src/main/java/com/kropsz/github/backendboxboxd/service/search/TeamService.java
@@ -1,4 +1,4 @@
-package com.kropsz.github.backendboxboxd.service;
+package com.kropsz.github.backendboxboxd.service.search;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -7,21 +7,23 @@ import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.kropsz.github.backendboxboxd.entities.Driver;
-import com.kropsz.github.backendboxboxd.repository.DriverRepository;
+import com.kropsz.github.backendboxboxd.entities.Team;
+import com.kropsz.github.backendboxboxd.repository.TeamRepository;
 
 import lombok.RequiredArgsConstructor;
 
 @Service
 @RequiredArgsConstructor
-public class DriverService {
+public class TeamService{
     
-    private final DriverRepository driverRepository;
+    private final TeamRepository teamRepository;
 
-     public static Specification<Driver> hasProperty(String property, String value) {
-        return (driver, cq, cb) -> cb.equal(driver.get(property), value);
+    
+    public static Specification<Team> hasProperty(String property, String value) {
+        return (team, cq, cb) -> cb.equal(team.get(property), value);
     }
 
+    
     public static Sort orderByProperty(String property, Sort.Direction direction) {
         if (direction == null) {
             direction = Sort.Direction.ASC;
@@ -32,12 +34,14 @@ public class DriverService {
         return Sort.by(direction, property);
     }
 
+    
     @Transactional(readOnly = true)
-    public Page<Driver> getDriversByProperty(String property, String value, Pageable pageable) {
+    public Page<Team> getTeamsByProperty(String property, String value, Pageable pageable) {
         if (property == null && value == null) {
-            return driverRepository.findAll(pageable);
+            return teamRepository.findAll(pageable);
         }
-        Specification<Driver> spec = hasProperty(property, value);
-        return driverRepository.findAll(spec, pageable);
+        Specification<Team> spec = hasProperty(property, value);
+        return teamRepository.findAll(spec, pageable);
     }
+
 }
